🗑️ fix: PR comments and split up Credits history section
