+++ src/components/organisms/CreditsHistorySection/CreditsHistorySection.spec.tsx
@@ -4,20 +4,6 @@ import CreditsHistorySection from '@/components/organisms/CreditsHistorySection/
 import { TestOvermindWrapper } from '@/test/TestOvermindWrapper'
 import { MockData } from '@/test/TestOvermindMockData'
 
-Object.defineProperty(window, `matchMedia`, {
-  writable: true,
-  value: jest.fn().mockImplementation((query) => ({
-    matches: false,
-    media: query,
-    onchange: null,
-    addListener: jest.fn(), // deprecated
-    removeListener: jest.fn(), // deprecated
-    addEventListener: jest.fn(),
-    removeEventListener: jest.fn(),
-    dispatchEvent: jest.fn(),
-  })),
-})
-
 describe(`CreditsHistorySection`, () => {
   test(`renders loading state when creditsHistoryData is undefined`, () => {
     const { container } = render(
