+++ src/components/organisms/CreditsHistorySection/CreditsHistorySectionSkeleton.spec.tsx
@@ -1,0 +1,13 @@
+import { render } from '@testing-library/react'
+
+import { TestOvermindWrapper } from '@/test/TestOvermindWrapper'
+import CreditsHistorySkeleton from '@/components/organisms/CreditsHistorySection/CreditsHistorySectionSkeleton'
+
+test(`renders loading state when creditsHistoryData is undefined`, () => {
+  const { container } = render(
+    <TestOvermindWrapper>
+      <CreditsHistorySkeleton />
+    </TestOvermindWrapper>,
+  )
+  expect(container).toMatchSnapshot()
+})
