+++ src/components/organisms/CreditsHistorySection/IssuedRetiredChartBody.spec.tsx
@@ -1,0 +1,13 @@
+import { render } from '@testing-library/react'
+
+import { TestOvermindWrapper } from '@/test/TestOvermindWrapper'
+import IssuedRetiredChartBody from '@/components/organisms/CreditsHistorySection/IssuedRetiredChartBody'
+
+test(`renders correctly`, () => {
+  const { container } = render(
+    <TestOvermindWrapper>
+      <IssuedRetiredChartBody />
+    </TestOvermindWrapper>,
+  )
+  expect(container).toMatchSnapshot()
+})
