+++ src/test/TestOvermindMockData.ts
@@ -79,30 +79,32 @@ const STATE_CARBON_FULL_REGION_COUNTRY_NO_HOVER: DataState = {
   carbonMapDataFiltered: CARBON_MAP_DATA_FILTERED,
 }
 
+const CREDIT_HISTORY_CHART_DATA = [
+  /* eslint-disable no-magic-numbers */
+  {
+    id: `issued`,
+    data: [
+      { x: new Date(2023, 3), y: 211306 },
+      { x: new Date(2023, 4), y: 30115653 },
+      { x: new Date(2023, 5), y: 110387057 },
+    ],
+  },
+  {
+    id: `retired`,
+    data: [
+      { x: new Date(2023, 3), y: 118885 },
+      { x: new Date(2023, 4), y: 26355689 },
+      { x: new Date(2023, 5), y: 25333735 },
+    ],
+  },
+  /* eslint-enable no-magic-numbers */
+]
+
 const CREDIT_HISTORY_DATA: CreditsHistoryDataState = {
   dataFilters: { region: SubRegion.LATIN_AMERICA_AND_THE_CARIBBEAN, timeframe: TimeframesData.MAX, country: `COL` },
   rawCreditsHistory: { data: CREDIT_HISTORY_MOCK_DATA },
   filteredCreditsHistory: {
-    /* eslint-disable no-magic-numbers */
-    chartData: [
-      {
-        id: `issued`,
-        data: [
-          { x: new Date(2023, 3), y: 211306 },
-          { x: new Date(2023, 4), y: 30115653 },
-          { x: new Date(2023, 5), y: 110387057 },
-        ],
-      },
-      {
-        id: `retired`,
-        data: [
-          { x: new Date(2023, 3), y: 118885 },
-          { x: new Date(2023, 4), y: 26355689 },
-          { x: new Date(2023, 5), y: 25333735 },
-        ],
-      },
-    ],
-    /* eslint-enable no-magic-numbers */
+    chartData: CREDIT_HISTORY_CHART_DATA,
     issued: 140714016,
     retired: 51808309,
   },
@@ -134,6 +136,7 @@ export const MockData = {
   STATE_CARBON_FULL,
   STATE_CARBON_FULL_REGION_COUNTRY,
   STATE_CARBON_FULL_REGION_COUNTRY_NO_HOVER,
+  CREDIT_HISTORY_CHART_DATA,
   CREDIT_HISTORY_DATA,
   CREDIT_HISTORY_DATA_EMPTY_CHART,
 }
